+++ app/src/main/java/com/rofwin/ContainerModel.kt
@@ -1,4 +1,4 @@
-package com.winlator
+package com.rofwin
 
 import kotlinx.serialization.Serializable